move arrow heads for better visibility
I suggest to have the arrow heads half-way to be able to see them, otherwise in real plots arrow heads overlap at the join points of the boxes and are indistinguishable. Basically done inefficiently by plotting lines first and then arrows that are shorter. Arrows are no simple R function to change because they seem to call on C code.